+++ src/App.jsx
@@ -4,7 +4,7 @@ import { LocalizationProvider } from "@mui/x-date-pickers/LocalizationProvider";
 import LandingPage from "./routes/landingPage/LandingPage";
 import About from "./routes/about/About";
 import AccountSettings from "./routes/accountSettings/AccountSettings";
-import AdminCalendars from "./routes/AdminCalendars";
+import AdminCalendars from "./routes/adminCalendars/AdminCalendars";
 import AdminPanel from "./routes/adminpanel/Adminpanel";
 import Calendar from "./routes/Calendar";
 import Calendars from "./routes/Calendars";
