+++ src/routes/Root.jsx
@@ -1,5 +1,5 @@
 import Header from "../components/header/Header";
-//import Footer from "../components/footer/Footer";
+import Footer from "../components/footer/Footer";
 import { Outlet } from "react-router-dom";
 export default function Root() {
   return (
@@ -8,7 +8,7 @@ export default function Root() {
       <main>
         <Outlet />
       </main>
-      {/*<Footer />*/}
+      {<Footer />}
     </>
   );
 }
