+++ src/routes/adminCalendars/adminCalendars.css
@@ -1,0 +1,22 @@
+.cards {
+  display: grid;
+  grid-template-columns: 33% 33% 33%;
+  grid-template-rows: 30vh 30vh;
+  column-gap: 10px;
+  row-gap: 15px;
+  justify-items: center;
+  align-items: center;
+  width: 100%;
+  place-items: center;
+  justify-content: center;
+}
+.dataCard {
+  height: 100%;
+  background-color: #eae0d7 !important;
+}
+
+#dropdown-item-button {
+  background-color: #ba6c2c !important;
+  color: #eae0d7 !important;
+  border-color: #ba6c2c !important;
+}
